Visual Studio builds: Fix Python and PERL handling
We may not be building the introspection files, so the build would likely fail
if we do not pass in PYTHON=... in the NMake command line, even if a Python
interpreter is in the PATH, unless introspection files are being built, because
we are running glib-mkenums, which will make use of Python (or, PERL, for older
versions).

Fix this by moving where we set up the PYTHON path so that it will be set up
for all builds.

Likewise, make sure we set up a default value for calling PERL, so that if it
is to be used, we can opt to pass in the full path of the PERL interpreter, or
make sure the PERL interpreter is indeed in the PATH.