Visual Studio builds: Support ARM64 Windows builds
This adds build support for Visual Studio builds to build ARM64 Windows
binaries, which can either be done on a normal x86/x86-64 Windows machine, or
directly on an ARM64 Windows machine, provided that the Visual Studio ARM64
(cross-)compiler is installed and activated.

Note that at this point there is no support for building introspection files as
there are currently no official Python ARM64 Windows binaries, which is needed
to build GObject-Introspection.