autotools: remove gtk-doc
Autotools will be replaced by Meson. During a short period of time there
will be both Autotools and Meson support to compile Amtk, to not break
CI, gnome-build-meta etc. But with Meson the files in docs/reference/
can be reorganized, to avoid hard-coding the API version in the filename
for example.

So it's easier to first remove the Autotools support for building the
API documentation, then implement the Meson support for building the
docs and changing at the same time the files in docs/reference/ to
better accomodate Meson. Then the full Autotools build system can be
removed from Amtk once the Meson migration is complete and works fine.