build: add gobject_introspection option
Thanks to Chun-wei Fan for an initial patch.

There are occasions where one may not wish to build the introspection
data, or is unable to do so, such as during a cross-compilation
scenario.

The option is just a boolean, I prefer to avoid automagic configuration
(have automatic code to decide whether to enable/disable GI, that would
be more subject to build system code changes).

Closes https://gitlab.gnome.org/GNOME/amtk/-/merge_requests/6