Do not set G_GNUC_CONST for get_type() functions
IIRC G_GNUC_CONST means something else: according to its API, we should
not use it for get_type() functions.

It's an optimization that happens to work, currently at least, and with
GCC. But there are no guarantees that it will still work in the distant
future.

And, the less code the better, it's slightly easier to understand.