build: don't specify any 'soversion' or 'version' for the library
Apparently Meson differs from Libtool wrt. current.revision.age version.

Amtk has normally not break its API and ABI since the 5.0.0 version. So
it's much easier to not mention any 'version' or 'soversion' argument to
the library() Meson function call.

In case an ABI break is necessary in the future, then a 'version' or
'soversion' argument can be re-added.