build: set 'soversion' to 0 in the library() call
The 5.6.0 release introduced an unintentional soname bump.
See also previous commit.

Specifying 0 is necessary to avoid a soname change compared to the 5.4.x
releases. That way it's easier to update the package in Linux
distributions.

'soversion' is easier to use than the 'version' argument. In the future,
if an ABI break is done, it will also be easier (just increment the
number).

So with this commit, on Linux the filename is now:
libamtk-5.so.0

with a symbolic link:
libamtk-5.so -> libamtk-5.so.0